backend/pyproject.toml
# pyproject.toml
[project]
name = "smartbuy-backend"
version = "0.1.0"
description = "SmartBuy Deal Aggregation Engine - Milestone 1"
authors = [{name = "SmartBuy Team", email = "team@smartbuy.com"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.35",
    "pydantic>=2.9.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.5.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
    "structlog>=24.4.0",
    "bcrypt>=4.2.0",
    "pyjwt>=2.9.0",
    "openai>=1.50.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "black>=24.8.0",
    "isort>=5.13.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

# Directory Structure:
# Octopus/backend/
# ├── app/
# │   ├── __init__.py
# │   ├── core/
# │   │   ├── __init__.py
# │   │   ├── config.py
# │   │   └── database.py
# │   ├── models/
# │   │   ├── __init__.py
# │   │   └── deal.py
# │   ├── schemas/
# │   │   ├── __init__.py
# │   │   └── deal.py
# │   ├── routers/
# │   │   ├── __init__.py
# │   │   └── deals.py
# │   └── services/
# │       ├── __init__.py
# │       └── deal_service.py
# ├── main.py
# ├── pyproject.toml
# └── README.md